Fix the new TPM support to allow Key principals.
This commit changes the new TPM scripts to work if a TPM isn't present.
In particular, it adds guards to only add a TPM principal if the
appropriate files are available. And it fixes a bug in the TPM support
in tao_admin.go to return the errors in the case of errors in finding
the TPM files.

+++ go/run/scripts/set_up_domain.sh
@@ -21,7 +21,7 @@ DOMAIN_PATH=$(mktemp -d /tmp/domain.XXXXXX)
 HOST_REL_PATH=linux_tao_host
 FAKE_PASS=BogusPass
 TPM="/dev/tpm0"
-PCRS="17,18" # PCR registers of TPM 
+PCRS="17,18" # PCR registers of TPM
 AIKBLOB="${HOME}/aikblob"
 
 # Create the linux_host key before creating the domain. This is needed in this
@@ -31,7 +31,7 @@ AIKBLOB="${HOME}/aikblob"
 "$ADMIN" -operation key -domain_path $DOMAIN_PATH -pass $FAKE_PASS \
 	-config_template "$TEMPLATE" $HOST_REL_PATH
 
-# Get key host name, add it to the template. 
+# Get key host name, add it to the template.
 KEY_NAME=$("$ADMIN" -config_template "$TEMPLATE" -domain_path $DOMAIN_PATH \
 	-pass $FAKE_PASS -logtostderr $HOST_REL_PATH)
 
@@ -56,31 +56,22 @@ echo host_name: \"$KEY_NAME\" >> $TEMP_FILE
 	-add_linux_host -add_guard -domain_path $DOMAIN_PATH -pass $FAKE_PASS \
 	-config_template $TEMP_FILE -logtostderr
 
-# Add TPM principal to domain. 
-sudo "$ADMIN" -operation policy -add_tpm \
- 	-principal tpm -tpm $TPM -pcrs $PCRS -aikblob $AIKBLOB \
-	-pass $FAKE_PASS -domain_path $DOMAIN_PATH \
-	-config_template $TEMP_FILE -logtostderr
+# Add TPM principal to domain, if one exists.
+if [ -f "$AIKBLOB" ] && [ -f "$TPM" ]; then
+  sudo "$ADMIN" -operation policy -add_tpm \
+          -principal tpm -tpm $TPM -pcrs $PCRS -aikblob $AIKBLOB \
+          -pass $FAKE_PASS -domain_path $DOMAIN_PATH \
+          -config_template $TEMP_FILE -logtostderr
 
-# Get the TPM host name, add it to the template. 
-TPM_NAME=$(sudo "$ADMIN" -operation principal \
- 	-principal tpm -tpm $TPM -pcrs $PCRS -aikblob $AIKBLOB \
-	-pass $FAKE_PASS -domain_path $DOMAIN_PATH \
-	-config_template $TEMP_FILE -logtostderr)
+  # Get the TPM host name, add it to the template.
+  TPM_NAME=$(sudo "$ADMIN" -operation principal \
+          -principal tpm -tpm $TPM -pcrs $PCRS -aikblob $AIKBLOB \
+          -pass $FAKE_PASS -domain_path $DOMAIN_PATH \
+          -config_template $TEMP_FILE -logtostderr)
 
-cat "$TEMPLATE" | sed "s/REPLACE_WITH_DOMAIN_GUARD_TYPE/$GUARD/g" > $TEMP_FILE
-echo host_name: \"$TPM_NAME\" >> $TEMP_FILE
-
-# Add trusted applications. 
-#"$ADMIN" -operation policy -add_programs -pass $FAKE_PASS \
-#	-domain_path $DOMAIN_PATH -config_template $TEMP_FILE \
-#	-canexecute $GOPATH/bin/demo_server \
-#	-logtostderr
-#
-#"$ADMIN" -operation policy -add_programs -pass $FAKE_PASS \
-#	-domain_path $DOMAIN_PATH -config_template $TEMP_FILE \
-#	-canexecute $GOPATH/bin/demo_client \
-#	-logtostderr
+  cat "$TEMPLATE" | sed "s/REPLACE_WITH_DOMAIN_GUARD_TYPE/$GUARD/g" > $TEMP_FILE
+  echo host_name: \"$TPM_NAME\" >> $TEMP_FILE
+fi
 
 rm $TEMP_FILE
 echo "Temp domain directory: $DOMAIN_PATH"
